Starts nextjs migration
