+++ .babelrc
@@ -1,0 +1,11 @@
+{
+  "plugins": [
+    [
+      "babel-plugin-styled-components",
+      {
+        "ssr": true
+      }
+    ]
+  ],
+  "presets": ["next/babel"]
+}
