+++ package.json
@@ -4,30 +4,22 @@
   "private": true,
   "dependencies": {
     "@material-ui/core": "^4.11.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
     "@types/recharts": "^1.8.15",
     "@types/styled-components": "^5.1.2",
     "axios": "^0.20.0",
-    "prop-types": "^15.7.2",
+    "next": "^10.2.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-ga": "^3.1.2",
-    "react-jss": "^10.3.0",
     "react-lottie": "^1.2.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.1",
     "react-spring": "^8.0.27",
-    "recharts": "^1.8.5",
     "renature": "^0.6.2",
     "styled-components": "^5.1.1"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "lint": "eslint --ext js,jsx,ts,tsx .",
     "format": "prettier --write \"**/*.{ts,tsx,js,jsx,json}\""
   },
@@ -64,14 +56,13 @@
     ]
   },
   "devDependencies": {
-    "@types/jest": "^26.0.9",
     "@types/node": "^14.0.27",
     "@types/react": "^16.9.46",
     "@types/react-dom": "^16.9.8",
     "@types/react-lottie": "^1.2.5",
-    "@types/react-router-dom": "^5.1.5",
     "@typescript-eslint/eslint-plugin": "^3.9.0",
     "@typescript-eslint/parser": "^3.9.0",
+    "babel-plugin-styled-components": "^1.12.0",
     "eslint": "^6.6.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
@@ -80,6 +71,6 @@
     "husky": "^4.3.0",
     "lint-staged": "^10.3.0",
     "prettier": "^2.0.5",
-    "typescript": "^3.9.7"
+    "typescript": "^4.2.4"
   }
 }
